Renamed eval.py to chain.py
