+++ demo-src/demo.py
@@ -1,6 +1,5 @@
 import gradio as gr
-from eval import query_to_answer
-
+from chain import query_to_answer
 
 with gr.Blocks(title="TMLS Hackathon 2024") as demo:
     gr.Interface(fn=query_to_answer, inputs="textbox", outputs="textbox", allow_flagging="never")
